refactor: move code related to deployment log files into its own class
