+++ AMW_rest/src/main/java/ch/mobi/itc/mobiliar/rest/RESTApplication.java
@@ -23,6 +23,7 @@ package ch.mobi.itc.mobiliar.rest;
 import ch.mobi.itc.mobiliar.rest.Analyze.TestGenerationRest;
 import ch.mobi.itc.mobiliar.rest.auditview.AuditViewRest;
 import ch.mobi.itc.mobiliar.rest.deployments.DeploymentDtoCsvBodyWriter;
+import ch.mobi.itc.mobiliar.rest.deployments.DeploymentsLogRest;
 import ch.mobi.itc.mobiliar.rest.deployments.DeploymentsRest;
 import ch.mobi.itc.mobiliar.rest.environments.EnvironmentsRest;
 import ch.mobi.itc.mobiliar.rest.exceptions.*;
@@ -63,6 +64,7 @@ public class RESTApplication extends Application {
         resources.add(ResourceRelationTemplatesRest.class);
         resources.add(HostNamesRest.class);
         resources.add(DeploymentsRest.class);
+        resources.add(DeploymentsLogRest.class);
         resources.add(EnvironmentsRest.class);
         resources.add(AuditViewRest.class);
         resources.add(RestrictionsRest.class);
